cilium: encryption, clear mark from xfrm state in endpoint route modes
In endpoint route modes we have routes in the system to the endpoints so we
can skip the route back to the interface with bpf_network.o set, previously
used to do the redirect needed when the stack can not route.

This will work back to 4.14 kernels where XFRM_OUTPUT_MARK was added. If we
need to go back further an alternative route/mark/rule will need to be
added.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>